Make entityPrefabs use Vector2f for base creation

+++ SeaGame/src/EntityPrefabs.h
@@ -17,16 +17,16 @@ public:
 	static entity_ptr cannonBall(Game* g, std::weak_ptr<Entity> spawner, sf::Vector2f pos, float rot);
 	static entity_ptr enemyChasingShip(Game* g, sf::Vector2f pos);
 	static entity_ptr explosion(Game* g, sf::Vector2f pos);
-	static entity_ptr militaryBase(Game* g, sf::Vector2i pos);
-	static entity_ptr generationBase(Game* g, sf::Vector2i pos, GameResource res);
+	static entity_ptr militaryBase(Game* g, sf::Vector2f pos);
+	static entity_ptr generationBase(Game* g, sf::Vector2f pos, GameResource res);
 	static entity_ptr ferryShip(Game * g, sf::Vector2f pos, std::weak_ptr<Entity> from, std::weak_ptr<Entity> to);
-	static entity_ptr city(Game* g, sf::Vector2i pos);
-	static entity_ptr pirateBase(Game* g, sf::Vector2i pos);
+	static entity_ptr city(Game* g, sf::Vector2f pos);
+	static entity_ptr pirateBase(Game* g, sf::Vector2f pos);
 	static entity_ptr defensePirateShip(Game* g, sf::Vector2f pos, std::weak_ptr<Entity> base);
 	static entity_ptr attackPirateShip(Game* g, sf::Vector2f pos, std::weak_ptr<Entity> base, std::weak_ptr<Entity> target);
-	static entity_ptr pirateFortress(Game* g, sf::Vector2i pos);
-	static entity_ptr conversionBase(Game* g, sf::Vector2i pos, GameResource res);
-	static entity_ptr resourceSource(Game* g, sf::Vector2i pos, GameResource res);
+	static entity_ptr pirateFortress(Game* g, sf::Vector2f pos);
+	static entity_ptr conversionBase(Game* g, sf::Vector2f pos, GameResource res);
+	static entity_ptr resourceSource(Game* g, sf::Vector2f pos, GameResource res);
 	// Get the Box2d BodyDef and fixtures needed for specific entities
 	static b2Body* getShipBody(Game* g, sf::Vector2f pos, float rot);
 	static b2Body* getCannonballBody(Game* g, sf::Vector2f pos, float rot);
